fix: resolve console UI build and ESLint budget failures
Move chart color slug helper out of chart.tsx, drop invalid Rectangle
props in WidgetChart, and split Storybook helpers into fixture/decorator
modules to satisfy react-refresh and no-console budgets.
