+++ web_src/src/pages/app/console/NodesPanelCard.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 import type { ConsolePanel } from "@/hooks/useCanvasData";
 
 import { NodesPanelCard } from "./NodesPanelCard";
-import { MockConsoleProvider, PanelFrame } from "./__stories__/storyHelpers";
+import { MockConsoleProvider, PanelFrame } from "./__stories__/storyDecorators";
 
 /**
  * Multi-node ("Key Nodes") panel. Renders a compact list of canvas nodes with
